(unfinished) declutter, recenter/reanim options
added option for offset distance- implemented on a per-anim basis, offsets the initial popup location so as not to obstruct the player view
added option to recenter on hit- on hit, the popup is reposititioned, effectively as if the popup was recreated. this was the default behavior previously; the new default behavior is to change the popup text but not interrupt or restart its animation
added an option for popup alpha
(unfinished) added screen center declutter- when a popup is close to the center of the screen, it becomes transparent. there are two options to control the inner and outer boundary/radius for this declutter area. calculations for this are incomplete, and the menu options need to be refactored for consistency (the normal alpha setting should be integrated)